clean

+++ src/main/java/trie/BloomFilter.java
@@ -50,7 +50,7 @@ public class BloomFilter {
 				int h = 0;
 				for (int j = (i * 4); j < (i * 4) + 4; j++) {
 					h <<= 8;
-					h |= ((int) digest[j]) & 0xFF;
+					h |= (digest[j]) & 0xFF;
 				}
 				result[k] = h;
 			}
